ci: remove parnsips-emu symlink from node_modules in test workflow
This is necessary because pnpm creates a symlink, but
vite-plugin-wasm-pack wants to copy things... Perhaps this could
be resolved more elgantly by having vite-plugin-wasm-pack support
pnpm, but this works for now. Also, it's necessary to put the
`pnpm run build` before `pnpm run check` because the copying of
parsnips-emu to node_modules will only happen on `dev` or `build`
but it must be done before `check` so the type definitions are
there.