test!: pass native-endian memory instead of little-endian
This commit breaks all the tests on big-endian platforms again. This
commit is a step in the right direction when it comes to being
consistent about endianness. As far as I can tell, up until this point,
values in registers have been native-endian, but memory has always been
assumed to be little-endian.

From this point onwards, the plan is that everything will be
native-endian. At some point further down the road, support for
configurable endianness may be added so that the endianness opposite to
that of the host platform can be emulated.